% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacSynth.infer.R
\name{FacSynth.infer}
\alias{FacSynth.infer}
\title{Conduct statistical inference for factor model approach to estimate average treatment effect on treated.}
\usage{
FacSynth.infer(
  FacSynth.prep.obj = NULL,
  STNRY = FALSE,
  conf = 0.95,
  y1 = NULL,
  x1 = NULL,
  y2 = NULL,
  x2 = NULL,
  time.variable = NULL,
  time.prior = NULL,
  time.post = NULL
)
}
\arguments{
\item{y1}{A matrix that contains values of the treatment unit over the pre-intervention periods.}

\item{x1}{A matrix that contains values of the control units (plus an intercept) over the pre-intervention periods.}

\item{y2}{A matrix that contains values of the treatment unit over the post-intervention periods.}

\item{x2}{A matrix that contains values of the control units (plus an intercept) over the post-intervention periods.}

\item{time.variable}{A scalar identifying column number or column-name character string associated with period (time) data. The time variable has to be numeric.}

\item{time.prior}{The number of pre-treatment periods.}

\item{time.post}{The number of post-treatment periods.}

\item{Facsynth.prep.obj}{The object that comes from running \code{FacSynth.prep}. This objects contains all matrices and parameters necessary for running the function \code{FacSynth.infer}. Therefore, once users import this object, there is no need to specify any of matrices and parameters manually.}
}
\value{
A list which contains ATT derived from factor model approach, a series of confidence intervals, R-square and adjusted R-square.
\item{ATT Estimated by FM}{The value of ATT estimated by factor model.}
\item{Confidence Interval}{A matrix of intervals across all the confidence level(s) required.}
\item{R-squared}{The value of R-squared.}
\item{Adjusted R-squared}{The value of adjusted R-squared.}
}
\description{
This function provides users with a formal inference theory for using the factor model approach to estimate ATT, along with a new criterion proposed to more accurately select the number of factors.
}
\details{
To run this function, we need users to supply data object obtained by running the \code{FacSynth.prep}, or users should manually input relevant matrices and numeric values to fill in the corresponding arguments.
}
\examples{
  ##In order to conduct inference properly, we recommended users to directly use the list object
  ##derived from running FacSynth.prep() and determine the stationarity of the outcome.

  #load data from the package.
  data(synth.data)

  #extract relevant components necessary for running FacSynth.infer()
  #from wide-format panel data.
  FacSynth.prep.out =
    FacSynth.prep(
     data = synth.data,
     treatment.identifier = 9,
     controls.identifier = c(2:8),
     time.variable = 1,
     time.prior = c(1984:1989),
     time.post = c(1990:1998)
     )

  #then directly use this object to conduct formal inference process and then
  #obtain estimated ATT and its confidence intervals in different levels.
  FacsSynth.infer.out =
    FacSynth.infer(
     FacSynth.prep.out,
     STNRY = FALSE,
     conf = c(0.99, 0.95, 0.90)
     )
}
\references{
Li K, Sonnier G (2020) Statistical inference for factor model approach to estimate causal effects in quasi-experimental settings. Working Paper, University of Texas at Austin, Austin, TX.
}
