% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FacSynth.prep.R
\name{FacSynth.prep}
\alias{FacSynth.prep}
\title{Construct a list of matrices from panel dataset to be loaded into other functions needed.}
\usage{
FacSynth.prep(
  data = NULL,
  treatment.identifier = NULL,
  controls.identifier = NULL,
  time.variable = NULL,
  time.prior = NULL,
  time.post = NULL
)
}
\arguments{
\item{data}{The dataframe with required wide format.}

\item{treatment.identifier}{A scalar identifying the column number of treatment unit in the dataframe or a character string giving the column name of treatment unit in the dataframe.}

\item{time.variable}{A scalar identifying column number or column-name character string associated with period (time) data. The time variable has to be numeric.}

\item{time.prior}{A numeric vector identifying the row numbers corresponding to pre-treatment periods.}

\item{time.post}{A numeric vector identifying the row numbers corresponding to post-treatment periods.}

\item{control.identifier}{A scalar identifying the column numbers of control units in the dataframe or a vector of character strings giving the column names of control units in the dataframe.}
}
\value{
A list with a series of components prepared for running \code{FacSynth.infer}.
\item{Y1}{A matrix of treatment unit data for pre-treatment periods.}
\item{X1}{A matrix of control units data for pre-treatment periods.}
\item{Y2}{A matrix of treatment unit data for post-treatment periods.}
\item{X2}{A matrix of control units data for post-treatment periods.}
\item{time}{A matrix of period(time) data.}
\item{T}{A scalar identifying the column number of period(time) data.}
\item{T1}{A scalar identifying the number of pre-treatment periods.}
\item{T2}{A scalar identifying the number of post-treatment periods.}
\item{n}{A scalar identifying the column number of control units data.}
}
\description{
This function extracts relevant data objects from the given dataframe and produce a list of matrices necessary for running the factor model approach.
}
\details{
Users should import a dataframe ("data") with required wide format, identify the columns associated with treatment and control units respectively, time variable, the pre-treatment and post-treatment time periods.
}
\examples{
  ##First example: wide-format simulated dataset sourced from Package "gsynth".

  #load data from the package.
  data(simdata_wide)

  #extract relevant components necessary for running synth.meth()
  #from wide-format panel data.
  FacSynth.prep.out =
    FacSynth.prep(
     simdata_wide,
     treatment.identifier = 2,
     controls.identifier = c(7:50),
     time.variable = 1,
     time.prior = c(1:20),
     time.post = c(21:30)
     )

  ## Second example: The economic impact of terrorism in the
  ## Basque country using data from Abadie and Gardeazabal (2003)

  #load data from the package.
  data(basque)

  #extract relevant components necessary for running synth.meth()
  #from wide-format panel data.
  FacSynth.prep.out =
    FacSynth.prep(
     data = basque,
     treatment.identifier = 5,
     controls.identifier = c(2:4,6:19),
     time.variable = "year",
     time.prior = c(1964:1969),
     time.post = c(1970:1985)
     )

}
\references{
Li K, Sonnier G (2020) Statistical inference for factor model approach to estimate causal effects in quasi-experimental settings. Working Paper, University of Texas at Austin, Austin, TX.
}
